test(ApiAcceptanceTest): API e2e 테스트 클래스명 변경
