+++ src/test/java/roomescape/acceptance/ApiAcceptanceTest.java
@@ -14,17 +14,7 @@ import java.util.stream.Stream;
 import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
-public class ReservationAcceptanceTest implements AcceptanceTest {
-
-    @Test
-    @DisplayName("[Step2, Step5] 예약 목록을 조회한다.")
-    void getReservations() {
-        RestAssured.given().log().all()
-                .when().get("/reservations")
-                .then().log().all()
-                .statusCode(200)
-                .body("size()", is(0));
-    }
+public class ApiAcceptanceTest implements AcceptanceTest {
 
     @Test
     @DisplayName("[Step7] 예약 시간을 생성한다.")
@@ -41,8 +31,55 @@ public class ReservationAcceptanceTest implements AcceptanceTest {
                 .header("Location", "/times/1");
     }
 
+    @Test
+    @DisplayName("[Step7] 예약 시간 목록을 조회한다.")
+    void getReservationTimes() {
+        RestAssured.given().log().all()
+                .when().get("/times")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+
+
     @TestFactory
-    @DisplayName("[Step3, Step6] 예약을 추가한다.")
+    @DisplayName("[Step7] 예약 시간을 추가하고 삭제한다.")
+    Stream<DynamicTest> createThenDeleteReservationTime() {
+        return Stream.of(
+                dynamicTest("예약을 하나 생성한다.", this::createReservationTime),
+                dynamicTest("예약이 하나 생성된 예약 목록을 조회한다.", this::getReservationTimesWithSizeOne),
+                dynamicTest("예약을 하나 삭제한다.", this::deleteOneReservationTime),
+                dynamicTest("예약 목록을 조회한다.", this::getReservationTimes)
+        );
+    }
+
+    void getReservationTimesWithSizeOne() {
+        RestAssured.given().log().all()
+                .when().get("/times")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(1));
+    }
+
+    void deleteOneReservationTime() {
+        RestAssured.given().log().all()
+                .when().delete("/times/1")
+                .then().log().all()
+                .statusCode(204);
+    }
+
+    @Test
+    @DisplayName("[Step2, Step5] 예약 목록을 조회한다.")
+    void getReservations() {
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0));
+    }
+
+    @TestFactory
+    @DisplayName("[Step3, Step6, Step8] 예약을 추가한다.")
     Stream<DynamicTest> createReservation() {
         return Stream.of(
                 dynamicTest("예약 시간을 생성한다.", this::createReservationTime),
@@ -51,7 +88,7 @@ public class ReservationAcceptanceTest implements AcceptanceTest {
     }
 
     @TestFactory
-    @DisplayName("[Step3, Step6] 예약을 추가하고 삭제한다.")
+    @DisplayName("[Step3, Step6, Step8] 예약을 추가하고 삭제한다.")
     Stream<DynamicTest> createThenDeleteReservation() {
         return Stream.of(
                 dynamicTest("예약 시간을 생성한다.", this::createReservationTime),
